requirements.txt
vkbottle
pillow
aiofiles